requirements.txt
PyGithub
slacker
jenkinsapi